Move AlertSidebar and related components to separate file
